Move google tools into separate file, move tools into general file in tools folder
