+++ local_operator/tools/__init__.py
@@ -1,0 +1,7 @@
+"""
+This module provides a collection of utility tools for the local_operator framework, including
+web browsing, file system indexing, web search, image generation and alteration, credential
+management, scheduling, and email sending functionalities. The tools are designed to be
+extensible and integrate with various external APIs and services, such as Radient, FAL, SERP API,
+Tavily, and Google services.
+"""
